Refac: Adding user collection type to firestore
A @types folder and UserProps file was created to abstract the types from userContext provider. To be able to set the user types to the document created in firestore, it was used an utility function converter, also abstract from the user context provider, located in utils folder at typesConverter file.