Finish: create new users collections (planners and notes)
Three utils funtions was created to provide a better understanding of what is happening. A createStandardPlanners will create 3 planners options for a new user, will a planner sample. The createStandardNotes will provide a session notes for each split scheduled. As a refactior, a createNewUserStandarDocs will unify the usability of these fuctions into the userContext provider.